major additions to install-arch.md and qemu-create-windows.md
the 'install Arch/Artix' guide now has the same formatting and updates that the 'install Debian' guide received earlier this summer. also included in this update is a section on how to use 'connmanctl' to get wireless internet in an 'Artix Linux + runit' ISO environment.

the 'create Windows VMs in QEMU' guide now has a section on how to troubleshoot a Windows 11 install.

in addition, the 'install Debian' guide received some minor tweaks and edits that were included in the 'install Arch/Artix' guide.