FEATURE: added the first entries to the "programming syntax guides"
in review of all of the Advent of Code that was completed in 2025, a set of notes was written to document what was learned.

in this commit, notes have been added for Bash and for Python.

these notes were compiled over a 4.5 hour period, starting the evening of Dec 17 (Wed) and ending the evening of Dec 19 (Fri).